Faster Aysnc calculation of Probes
